backend/rag_service/requirements.txt
# RAG Service Dependencies

# Core
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Vector DB
qdrant-client>=1.7.0

# LLM Providers
google-generativeai>=0.3.0
httpx>=0.25.0
openai>=1.0.0

# NLP & Arabic
sentence-transformers>=2.2.0
pyarabic>=0.6.15

# Utilities
tenacity>=8.2.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# CLI & UI
typer>=0.9.0
rich>=13.0.0
tqdm>=4.65.0
